Update patch set 1
Patch Set 1:

So so as every operator gets picked up we have to add every operator name to operator.xml, that i think is a lot of job adding every operator detail to operator.xml. I would prefer that customization of operator name to be linked with an app/ an activity in CMparts so that one can customize the name from there without the need of adding every operator name to operator.xml, in such a way the operator name can be customized without any long steps.

Patch-set: 1