Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Tattoo:

The files should be removed, there's not tattoo product in our tree. If you are or know the maintainer contact us in order to add it, but as this change goes, it should not be part of it.

Dictionaries:

We've been through this in the past and there is a semi-working solution that downloads dictionaries from the market since these dictionaries take up a lot of space, maybe it's time to finish that work.

Patch-set: 1