Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Is there a compelling reason not to use the shipping 2.2 build fingerprint? You can still use the build version as 2.3.1, and if market apps are stupid and use the fingerprint instead of the OS version, it's better to call them out on it IMHO. Otherwise, this just causes problems with Google's stats and might cause apps to be unavailable.

Patch-set: 1