Update patch set 1
Patch Set 1: Verified

I've changed all targets except for DS and ONE to inherit from the new common_full makefile. This will exclude the built-in themes (but keep the theme engine and chooser) and the extra ringtones.

I'm not sure if trimming CMWallpapers is worth it, it's up to you guys.

I've tested this on a full build of Supersonic so far and everything seems fine.

Patch-set: 1