Update patch set 1
Patch Set 1:

I can make it get update binary from otapackage but update script would be a real hassle. Also update script won't be updated until we add a new binary n set its permission. That is pretty rare where we add a new binary. I can keep it updated with the new binaries whenever they are added or we have to wait for the coming weekend when I get free time to write the new script code in the leoupdate so it greps the autogenerated update script for permissions

Patch-set: 1