Update patch set 5
Patch Set 5:

Hi guys, just had to pop in here to give some advice. :)
I see that the device config is just copied from the buzz one. First off, the prebuilt kernel line can be removed, as it's redundant there. Then the product model should be changed to HTC Tattoo, which is what it is in original HTC rom. And the extra RIL settings? Those are not in Kalims and my configs, so I guess they're taken from another device config. Better remove them so they don't break anything on the Tattoo. The Windows Media thing can also be removed, as it's not support in Gingerbread. The inherit of themes_common is already being inherited via common_full, so that can also be deleted.

Patch-set: 5