Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Yeah, this is a fail :\
Our current apns-conf already has 133 duplicate entries, this adds another 10.

In the case of a duplicate, the first entry that matches the mcc+mnc+type combo wins; if that entry happens to be wrong (even if a correct one appears later on), the user gets no service. That's happening with my carrier and the list you're commiting, so it's a no go from me ;) I know the correct value for that one (it's probably the fix Ricardo Pinho was mentioning), but not for the other 142.

Take a look at http://pastebin.com/vtSYNtu5

Patch-set: 1