Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I know at least the 24001 values are duplicates because they actually should be. Telia owns and runs Halebop as a more youth-centered alternative to itself.

If one of them should be removed it should probably be the Halebop one as Telia is much more widely used according to the stats I've heard (I know someone working there). It's also more common for professional users from what I have seen.

It would be nice if Android supported a way to simply pick the one that fits if there are duplicates.

Patch-set: 1