LDPI: Remove carrier label from notification pulldown
This patch removes the carrier label and moves the clear
button to the end of the notifications list for LDPI devices.
We have very less screen real-estate as it is, and the carrier
label is sort of useless and takes up over 13% of screen height.

Maybe in the future the carrier label could be made an option for
all devices.

Change-Id: I72fc14cc7713d56b7817075e0527c09db032d29f