Update patch set 6
Patch Set 6:

Remove the notification power widget icons. It looks better with the mdpi ones. \n
USB Mass storage screen is not correct. \n
The CyanogenMod icon in settings should be same size as the others. \n\n

There's also an issue with that the initial language locale is set to "Ld (United States)", which gives incorrect keyboard layout, unless manually changing locale to a valid one. This happens because of adding the mdpi to PRODUCT_LOCALES += ldpi mdpi, and since mdpi is already the default, something goes wrong when the locale list is generated. Probably not much one can do about it, since this bug is there also when only defining mdpi, and it has been there since gingerbread came out.

Patch-set: 6