Update patch set 2
Patch Set 2:

But ok, if this is hard to understand why, do the following and you'll see why. Leave the definition in buzz.mk (or click.mk), build the rom, wipe and install the rom, go to settings and check default language. What is it set to? Now open keyboard, is it missing some of the numbers? Now remove the definition from buzz.mk (or click.mk), and add it to cyanogen_buzz.mk (or cyanogen_click.mk), build the rom and do the same routine as earlier. See anything different?

Patch-set: 2