Update patch set 1
Patch Set 1:

Arne,we have to decide an equal way to work on this, if we add only the diferent values, we do the same for all the files, or we mantain the full files at all the files. Since now most of the overlays are complete.

In fact I think mantaining complete files let us see better when original files changes or have new values since the developer has done the overlay.

Patch-set: 1