Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Lord_Mike in #cyanogenmod created the xml dictionary this is based on. He is also testing it currently but it worked fine last I heard.

I have not heard anymore about how we decided to handle the added dictionaries and the space issue on certain devices so I submitted it as usual.

Patch-set: 1