Allow builds to use Launcher2 instead of ADW
ADW can be slow on some devices. On the other hand, Launcher2 works fine
and provides a better "out-of-the-box" experience on them. This patch adds
a USE_ADW_LAUNCHER variable. When set to false, the build will include
Launcher2. When not set, or set to true, it will use ADW as usual.

This changeset also enables this new variable on zeppelin.

Change-Id: I4267a46bac93385475a0311f1826a823545daef3