Update patch set 1
Patch Set 1:

you mean an option to select Launcher2 or ADW? Because that would be great. I never really was impressed by the ADW as default launcher thing as i really like the stock experience...
if that's what you plan to do, wouldn't it be enough to just ship CM with both Launchers, so that the user can select its favorite Launcher on the first start?

Patch-set: 1