Update patch set 6
Patch Set 6: Looks good to me, but someone else must approve

Repo changes are already on my github and ready to fork for both device and props on a1002 and b9701. I just need to rebase with the new commits from today. I will work with the team tomorrow to hopefully complete the change. Are you good with the t8/smb_b9701 change as well?

Patch-set: 6