Update patch set 1
Patch Set 1:

The problem with removing ADW from /system is that unless the user installed ADW to /data before upgrading to a build without ADW he/she would loose their Homescreen setup and settings as far as I know since Android would purge it because the app is missing. At least that's what happened to me when I upgraded to CM6 back in the day without thinking.

With that said, I also enjoy the stock experiance or a slightly tweaked stock experiance over ADW. Too bad launcher Pro isn't open source since that provides just that but with the 100% smoothness Launcher2 lacks.

Patch-set: 1