LDPI: Fixed broken buttons in GMail and Browser
After searching, testing and compiling nearly two days I finally
found a way arround that issue.
(http://code.google.com/p/cyanogenmod/issues/detail?id=3587)

I used the mdpi images and only changed the height and the
nine-patch-bars, but not the width.
That seems to do the trick and all buttons look good as
far as I can see on my wildfire.
Packed the images to the overlay instead of editing the theme.

Please test on your ldpi devices, especially with apps
that use buttons with default style.

Change-Id: Ic48c6c44d23d0e6c8a6c3822f17fda653796ea62