Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

CM does not customize the at the device level any more than required.  

This should be submitted in a manner that impacts ALL devices, with the default being OFF.

Personally, I think it's a feature that CM does not have noisy bootups.

Patch-set: 1