Update patch set 1
Patch Set 1:

lolol this is actually my code I did a while ago, Keyan/Steve, would you be interested in me pushing a patch to at least allow CM to *support* a boot sound, but include none? or skip it?

I was on the fence about this considering I hate boot sounds as well.

Patch-set: 1