Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

Sorry for the confusion. Wrote the information about patch set 2 in the commit message (learning by doing mistakes it seems) so I had to make a patch set 3 just to restore the commit message.

I missed adding the lunch combo in vendorsetup in patch set 1 which is why the extra patch set was uploaded. (set 2 and 3 are identical for reasons explained above).

Patch-set: 3