Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

This is basically a blade with a 7 inch screen, so a lot of the work that has been done on the blade is directly transferable to the v9. 

I also have a v9 so can help out with maintaining. As far as I know everything works using kallt kaffe's config, with basically the same minor bugs the blade still has and possibly a few new bugs related to screen res/dpi.

I haven't done a build of this for a while, but it looked good last time I did.

Patch-set: 3