Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

Patch set 4: Added myself and Tom as maintainers in CHANGELOG.mkdn

As an answer to what is broken: Almost everything works very well, much thanks to the hard work Tom and Jacob have done with the ZTE Blade. I also reused some overlays from the Nook Color (encore) project to overcome some of the MDPI on WVGA issues.

Known issues:
- On first boot (after wipe) the lockscreen and/or SIM lock usually is shown in a HVGA "box" but I've never seen it happen again after that.

- Recent apps window uses full screen directly after boot but switches to a HVGA "box" after some use. I suppose this should affect all WVGA(480x800) devices running MDPI(160) but I haven't seen any bug reports on it. I've previously worked around the problem but with this change in frameworks/base: https://github.com/kalltkaffe/android_frameworks_base/commit/000ec061d43ab873b6d876b1e6931be8c283ec94
Not sure if that is an acceptable workaround or if the real issue should be investigated and fixed elsewhere (seems like a problem with MATCH_PARENT)

-Screen calibration is currently hardcoded which isn't ideal for a resistive screen but the generic calibration support I added for review erlier (http://review.cyanogenmod.com/#change,6724) did not catch any interest so I'll try to make a device specific calibration app that works directly with the kernel instead. !!This is propably the biggest limation right now!!

-I don't think forcing 2G or 3G works and I think this is the same with the ZTE Blade. It's stuck on auto.

According to http://stats.cyanogenmod.com/ there are 1107 installs of the unofficial V9 build at the time of writing.

Patch-set: 4