Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

why are you changing the makefile on me? #1 this board does NOT have all the hardware capabilities for FM which is exactly this was removed a while ago. #2 you are adding Torch which is called on elsewhere. #3 you added GanOptimizer when this phone doesn't even have a wifi calling app. #4 the kernel is prebuilt.

Patch-set: 2