Update patch set 2
Patch Set 2:

Jared;

Chill. I agree that the second revision is pointless and I'll merge patchset 1. However, and FYI, authorship and commitership are separate and distinct concepts in git. Resubmission of a patchset by a second (and 3rd, and 4th) person happens regularly (for merge conflicts, for example) and has no impact on authorship history.

Also... Where are the props

Patch-set: 2