Update patch set 4
Patch Set 4:

Is it ok to make changes to the device tree since it hasn't yet been forked?  
We found an error in power_profile.xml that is the reason we've had less than par battery life.  I've fixed it in my unofficial builds and would like to include it before the tree gets forked.

Patch-set: 4