Update patch set 3
Patch Set 3:

Corrected CHANGELOG.mkdn.

I had a chat with SuperKid and he would like to be on the maintainers list. Please note this was an incorrect assumption on my part, not a malicious move to take all the credit =)

Thanks again to everyone reviewing this code, I hope it meets the high standard set by the rest of CyanogenMod.

Patch-set: 3