Update patch set 3
Patch Set 3:

Currently we're using the AOSP branch of the Tiamat kernel by Cayniarb. This in turn is forked from the CM MSM7x30 repo. The AOSP branch has none of the "Tiamat" modifications.

I had planned to try and bring those changes into the CM MSM7x30 git repo in the next few days. As and when I succeed in this I will submit them as requested.

Patch-set: 3