Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

So this is just a general cleanup & standardization of the products files. This should make it easy to make changes to the products files the affect all devices in the future. And with the release of 2.3.6 (GRK36C), I also updated the BUILD_DISPLAY_ID too with this patch, since it came out while I was cleaning these up.

Patch-set: 1