Update patch set 1
Patch Set 1:

Got no problems with zram0 device being created on the Wildfire and the Tattoo, but I have it built into the kernel, and not as a module. I see you pushed a lot of unneeded zram code on your github, so I'm guessing you didn't get the zram source from mainline? There's only a few files needed. You can try the version from my github if you want, and see if the device it created ok.

Patch-set: 1