Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

z71 has a major WiFi bug, activates and connects fine for the first time,but if we turn it off and on again, it goes into loop trying to connect but fails, hanging settings.apk. Reboot is needed to correct the issue.

Patch-set: 1