Update patch set 2
Patch Set 2:

Ok last message promise :-), after racking my brain and my plethora of devices with this problem today I have come to the answer as to why it works for me and EVERY device i have yet some users are reporting not being able to connect. It has nothing to do with the device, i attempted using my sisters phone first with my sim than with hers. It has to do with my account having the "4g" (even though we know its fake on ATT) enabled on my account and with my sim because of this all of my H+ devices including the captivate can use the apn provided here, how ever if you have not purchased a new "4g" device, got a new sim or called att and had them enable this apn on your account you can not connect as of yet.  The problem is reverting to old apn is not the answer either as i can say that my devices and service suffer from using the WAP APN , any how to anyone that needs to know more feel free to contact me, otherwise Steve I think the only solution is to provide both (tech all 3 as new apn uses separate apn's for mms) if you wish i can submit a new revised patch with both new and old, otherwise im sure you can handle on your own.  This is important however as new devices merge in they will all be using this new apn so it is important, even if we choose to set default to WAP for now

Patch-set: 2