Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Is there a way we can build the themes on LDPI but not have them set as the default? In my experience, the themes work fine but some aspects don't look so good. That doesn't bother a lot of people, and I suspect people won't be too happy with them being removed altogether.

I completely agree with not having CyanBread as default for LDPI, but if possible I think we should still include the themes in the build.

The other alternative, of course, is to update the themes for full LDPI support.

Patch-set: 2