Update patch set 1
Patch Set 1:

There's still some work to be done... I'm happy to wait and polish up our device tree in the mean time.  I'm helping with this port.

We'll also need to remove the device/huawei/ideos tree, as it is actually the same device as the Huawei U8150 and conflicts with this one.  This device was initially marketed as the "IDEOS", but Huawei has since rebranded that into an entire brand/line).  The technical name for this device is the IDEOS U8150.

As far as I know it was koush who merged the ideos tree earlier in 2011, but nobody picked it up...

Patch-set: 1