Update patch set 2
Patch Set 2:

I started to convert some images to ldpi, but then I discovered that the statusbar icons I was converting are not as good quality as the default ones. They're not as clear and crisp. Also the battery status icons are not correct, as some of them has different sizes. So for now I think we'll stick with the default theme for ldpi.

Patch-set: 2