Update patch set 2
Patch Set 2:

Arne: That's probably because Google screwed up the drawables for mdpi in Gingerbread. I submitted a fix for the following files to CM and Google. 

My Google submission was 39 images and a few more was merged into CM but some were really hard to fix with just the HDPI drawables as a base, even with lots of manual touches. 

List of files: http://pastebin.com/DNsSB6ir

The many of the battery images have been created since then though when Google increased the number of them.

Patch-set: 2