Update patch set 2
Patch Set 2:

Yeah, the mdpi statusbar icons in cyanbread looks as bad as they were in the default theme, before your makeover of them. Also take a look at the 40% and 60% battery indicator images in cyanbread. They're not showing the correct level.

It would probably be a good idea to replace them with the ones from the default theme, after "blueifying" them.

Patch-set: 2