Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Guys, stop commenting.

It's that that we don't want to merge the device, but this needs to be updated as advised by Diogo before it can be merged.

Saying "I think you should merge this" doesn't speed it up.

Patch-set: 1