Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

If build time is an issue perhaps using an environment variable (eg NO_PNGCRUSH) could be used to change the default.
Otherwise this looks fine to me.
You don't need semicolons at the end of the line when the 'then' is on another line.

Patch-set: 1