Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Tanguy,

for PRODUCT_MANUFACTURER, I was using what (at least) sholes and droid2 are using.

For the fingerprint, I agree the old version was more 'correct' as it describes our build. But remember, the market isn't licensed for our builds, it's licensed for the stock builds, which is why we need to use stock fingerprints.

Some feedback regarding the issue:

http://forum.xda-developers.com/showpost.php?p=19150827&postcount=13403

http://forum.xda-developers.com/showpost.php?p=19113376&postcount=245

http://www.android-hilfe.de/custom-roms-fuer-motorola-defy/164384-cm7-market-problem-mit-kompatibilitaet.html#post2220980 (German)

Patch-set: 1