Fix Galaxy S 1 devices market filtering
This is a "backport" from changes made during ICS development.
Updates all the devices to show applications properly in the
android market through use of TARGET_DEVICE trickery. This is
probably going to mess with CMStats, but seeing as to how we're
moving onto ICS anyway, that doesn't seem like it'll be that
big of an issue. <device>mtd wasn't a real device and some
protected apps weren't showing up (such as anything by GameLoft).

Change-Id: Id644470337796a8d86dd4c1479ef4cce0216dbe1