Update patch set 1
Patch Set 1: Do not submit

This can't be merged without changes to the mirroring system, since it currently requires ro.product.device to be the exact same as the lunch combo.
It'll also break CMStats, and some other minor stuff. We may need to rethink our usage of identifier properties (probably create a distinct ro.cm.productid to store the lunch combo and use that as index for other stuff), but until then this can't be merged.

Patch-set: 1