Update patch set 1
Patch Set 1:

This doesn't break the build process when used as a build.prop override. This is simply for market filtering (which we have been successfully using on ICS for a while now). On top of that, the device zips still come out as "captivatemtd". Last but not least, the stats will be the only "broken" change, but we've already made the change for ICS and there are no more 7.x builds to be released any time soon (is 7.2 even still happening?)

Patch-set: 1