Update patch set 3
Patch Set 3:

Do you think you would be able to remove patchset3 from this commit, possibly move it into a new one to get pushed through while we are waiting for this. The build fingerprint is bogus and we should get that pushed through as soon as possible, could be causing some of the lockups while using the market. After I made this change I have not had the market lockup on me yet. Thanks for your consideration.

Patch-set: 3