Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I still haven't merged some olympus device stuff, I'll do soon. What worries me most though is the fingerprint unlock code; what are the odds of merging it in time? If there's not a chance I'd consider releasing it without support for the functionality, or at least releasing the RCs without a final release.

On the other hand, zeppelin should be good to go, and I bet the same applies for morrison and motus

Patch-set: 1