Update patch set 2
Patch Set 2:

Ricardo: since I don't have the device anymore, I'll have to look and see if someone else (from XDA or whereever) would be interested in maintaining the Sunfire, which is extremely close to the Olympus.

Will: if I'm not mistaken, this part of the tree was nearly cloned off of the pyramid overlay. It might be a remnant of the old way of using resources, I'd have to check.

Patch-set: 2