Update patch set 13
Patch Set 13:

This is not a matter of appreciation or user demand, as you know gerrit is not a user forum. 

That being said, this device will not be added to the gingerbread branch, not because we hate users but because there are technical reasons that don't allow it. You can expect it in the ics branch when it's ready.

Patch-set: 13