Update patch set 19
Patch Set 19: I would prefer that you didn't submit this

I apologise for the botched patch-set. (The diff and title were lost in the latest 2 patches)

I cherry picked patch 17, and changed and committed but it replaced the title and diff.

Can someone explain how to just change a commit but keep the diff?

Sorry again.

Patch-set: 19