Update patch set 2
Patch Set 2:

I'd like to help review this.

I've just acquired a ChaCha and  had independently ported a Clockwork recovery image to the device, with the hope of porting CM entirely. What's the best way to complete a review of this patch set for someone who has never hacked on CM before this review?

Patch-set: 2