Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

This is my first review of a port to a new device. It was a bit rough to get it going but I did the following steps:
http://pastebin.com/zqNFgGcA

I re-flashed my ChaCha recovery image with fastboot and can confirm that the freshly built recovery image works well enough to flash update-cm-7.2.0-RC0-Chacha-KANG-signed.zip - I booted quite quickly into the build.

Here's the logcat of the first boot: http://pastebin.com/p9xEiU9e

It appears to work reasonably well. I've also built the kernel located at https://github.com/adumont/htc-kernel-msm7227.git and I haven't tested building that kernel into this image, I figured it was out of scope for this review.

I'd be happy to confirm the hashes of the proprietary bits as I grabbed images of the original firmware before re-flashing it.

Patch-set: 2