Update patch set 3
Patch Set 3:

Harry,

I'm still doing some work on the g5. I ported the 2.6.35 kernel, which works well for me, but some users are complaining of issues that I'm not yet sure are genuinely kernel problems. I'm making sure that it works well enough to apply to be used, but if not, I will review the 2.6.32 kernel source to be sure that's running stable.

With or without the 2.6.35 kernel, I've also got some small fixes and updated device settings app to add to the gingerbread branch on the g5.

Patch-set: 3