Update patch set 1
Patch Set 1:

The BUILD_FINGERPRINT has had that inc/inc/inc forever, apparently.  Any reason to change it?

For the rest, matching the latest OTA versions, meh.  I don't think it really matters (do explain to me if it does), but it doesn't hurt either.  I'll give my +1.

Need to clean up the commit message though, so it's not so... meta. :)  FWIW, when I have questions and comments about my own patch, I'd put that in a self-review, so if the concerns are mooted there's no need for a patch set 2.

Patch-set: 1